Revert "Replace a Few IntFlags with Uint64Flags (#9959)" (#10163)
* Revert "Replace a Few IntFlags with Uint64Flags (#9959)"

This reverts commit 790bf031234f5bbd734fdb23f892e1041d293e4c.

* fix

* fix

+++ beacon-chain/sync/initial-sync/service.go
@@ -163,13 +163,13 @@ func (s *Service) Resync() error {
 }
 
 func (s *Service) waitForMinimumPeers() {
-	required := uint64(params.BeaconConfig().MaxPeersToSync)
+	required := params.BeaconConfig().MaxPeersToSync
 	if flags.Get().MinimumSyncPeers < required {
 		required = flags.Get().MinimumSyncPeers
 	}
 	for {
 		_, peers := s.cfg.P2P.Peers().BestNonFinalized(flags.Get().MinimumSyncPeers, s.cfg.Chain.FinalizedCheckpt().Epoch)
-		if uint64(len(peers)) >= required {
+		if len(peers) >= required {
 			break
 		}
 		log.WithFields(logrus.Fields{
